requirements.txt
streamlit==1.32.2
torch==1.13.1
torchvision==0.14.1
easyocr==1.6.2
opencv-python-headless==4.5.5.64
numpy
pandas
Pillow
